requirements.txt
boto3==1.33.13
openpyxl==3.1.2
python-pptx==0.6.23